refactor(token-info): split info into collapsible blocks
